+++ EstimoteIndoorLocationSDK/Headers/EILPositionedBeacon.h
@@ -5,6 +5,8 @@
 
 @class EILOrientedPoint;
 
+NS_ASSUME_NONNULL_BEGIN
+
 /** Represents a beacon with additional information about its position. Object is immutable. */
 @interface EILPositionedBeacon : NSObject
 
@@ -15,16 +17,18 @@
 
 /** Oriented point representing beacons position and orientation. */
 @property (nonatomic, strong, readonly) EILOrientedPoint *position;
+/** Beacons identifier. */
+@property (nonatomic, strong, readonly) NSString *identifier;
 /** Beacons mac address. */
-@property (nonatomic, strong, readonly) NSString *macAddress;
+@property (nonatomic, strong, readonly) NSString *macAddress __attribute ((deprecated(("Use identifier instead."))));
 /** Beacons color. */
 @property (nonatomic, assign, readonly) ESTColor color;
 /** Beacons proximity UUID. Can be nil, if unknown. */
-@property (nonatomic, strong, readonly) NSUUID *proximityUUID;
+@property (nonatomic, strong, readonly, nullable) NSUUID *proximityUUID;
 /** Beacons major value. Can be nil, if unknown. */
-@property (nonatomic, strong, readonly) NSNumber *major;
+@property (nonatomic, strong, readonly, nullable) NSNumber *major;
 /** Beacons minor value. Can be nil, if unknown. */
-@property (nonatomic, strong, readonly) NSNumber *minor;
+@property (nonatomic, strong, readonly, nullable) NSNumber *minor;
 
 
 #pragma mark Creating and Initializing Beacons
@@ -35,12 +39,55 @@
 /**
  * Returns a new beacon object.
  *
+ * @param identifier Beacon identifier (MAC address or identifier).
+ * @param position Beacon position.
+ * @return A positioned beacon initialized with identifier and position.
+ */
+- (instancetype)initWithBeaconIdentifier:(NSString *)identifier
+                                position:(EILOrientedPoint *)position;
+
+/**
+ * Designated initializer. Returns a new beacon object.
+ *
+ * @param identifier Beacon identifier (MAC address or identifier).
+ * @param position Beacon position.
+ * @param color Beacon color
+ * @return A positioned beacon initialized with identifier and position.
+ */
+- (instancetype)initWithBeaconIdentifier:(NSString *)identifier
+                                position:(EILOrientedPoint *)position
+                                   color:(ESTColor)color;
+
+/**
+ * Designated initializer. Returns a new beacon object.
+ *
+ * @param identifier Beacon identifier (MAC address or identifier).
+ * @param position Beacon position.
+ * @param color Beacon color
+ * @param proximityUUID Beacon proximity UUID.
+ * @param major Beacon major value.
+ * @param minor Beacon minor value.
+ * @return A positioned beacon initialized with provided data.
+ */
+- (instancetype)initWithBeaconIdentifier:(NSString *)identifier
+                                position:(EILOrientedPoint *)position
+                                   color:(ESTColor)color
+                           proximityUUID:(nullable NSUUID *)proximityUUID
+                                   major:(nullable NSNumber *)major
+                                   minor:(nullable NSNumber *)minor;
+
+/**
+ * Returns a new beacon object.
+ *
  * @param macAddress Beacon MAC address.
  * @param position Beacon position.
  * @return A positioned beacon initialized with MAC and position.
+ *
+ * This method is deprecated. Use initWithBeaconIdentifier:position: instead.
  */
 - (instancetype)initWithBeaconIdentifiedByMac:(NSString *)macAddress
-                                 withPosition:(EILOrientedPoint *)position;
+                                 withPosition:(EILOrientedPoint *)position
+                                 __deprecated_msg("Use initWithBeaconIdentifier:position: instead.");
 
 /**
  * Designated initializer. Returns a new beacon object.
@@ -49,10 +96,13 @@
  * @param position Beacon position.
  * @param color Beacon color
  * @return A positioned beacon initialized with MAC and position.
+ *
+ * This method is deprecated. Use initWithBeaconIdentifier:position:color: instead.
  */
 - (instancetype)initWithBeaconIdentifiedByMac:(NSString *)macAddress
                                  withPosition:(EILOrientedPoint *)position
-                                        color:(ESTColor)color;
+                                        color:(ESTColor)color
+                                        __deprecated_msg("Use initWithBeaconIdentifier:position:color: instead.");
 
 /**
  * Designated initializer. Returns a new beacon object.
@@ -64,13 +114,16 @@
  * @param major Beacon major value.
  * @param minor Beacon minor value.
  * @return A positioned beacon initialized with provided data.
+ *
+ * This method is deprecated. Use initWithBeaconIdentifier:position:color:proximityUUID:major:minor: instead.
  */
 - (instancetype)initWithBeaconIdentifiedByMac:(NSString *)macAddress
                                  withPosition:(EILOrientedPoint *)position
                                         color:(ESTColor)color
-                                proximityUUID:(NSUUID *)proximityUUID
-                                        major:(NSNumber *)major
-                                        minor:(NSNumber *)minor;
+                                proximityUUID:(nullable NSUUID *)proximityUUID
+                                        major:(nullable NSNumber *)major
+                                        minor:(nullable NSNumber *)minor
+                                        __deprecated_msg("Use initWithBeaconIdentifier:position:color:proximityUUID:major:minor: instead.");
 
 /**
  * Returns a new beacon object.
@@ -93,7 +146,7 @@
  * @param dict `NSDictionary` to deserialize from.
  * @return A positioned beacon created from `NSDictionary`.
  */
-+ (EILPositionedBeacon *)beaconFromDictionary:(NSDictionary *)dict;
++ (nullable EILPositionedBeacon *)beaconFromDictionary:(NSDictionary *)dict;
 
 /**
  * Serializes the positioned beacon to `NSDictionary`.
@@ -136,6 +189,8 @@
  *
  * @return A string representation of the EILPositionedBeacon.
  */
--(NSString *)description;
+- (NSString *)description;
 
 @end
+
+NS_ASSUME_NONNULL_END